Optimize Will Smith

+++ catscripts/Entertainment/Actors_and_actresses/Will_Smith/WillSmith.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-export KEYWORDS_ACTOR_WILLSMITH="Will(|ard)(| )Smith"
+KEYWORDS_ACTOR_WILLSMITH="Will(|ard)(| )Smith"
 
-egrep -i "$KEYWORDS_ACTOR_WILLSMITH" newpages.txt >> WillSmith.txt
-
-WILLSMITH=`stat --print=%s WillSmith.txt`
-
-if [ $WILLSMITH -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="WillSmith.txt"
-  export CATNAME="Will Smith"
-  $CATEGORIZE
-fi
 
-rm WillSmith.txt
+  WILLSMITH=`egrep -i "$KEYWORDS_ACTOR_WILLSMITH" newpages.txt`
+
+  if [ "$WILLSMITH" != "" ];
+  then
+    egrep -i "$KEYWORDS_ACTOR_WILLSMITH" newpages.txt > WillSmith.txt
+    export CATFILE="WillSmith.txt"
+    export CATNAME="Will Smith"
+    $CATEGORIZE
+    rm WillSmith.txt
+    unset WILLSMITH
+  fi
+
+fi
